refactor: Phase F revisions — Fraction Int64, Byron timestamp Int64, BIP32 UInt32
Three additions to scope after re-audit:

- Types/Fraction.swift: numerator/denominator Int → Int64. Used in
  governance/protocol ratios; the .uint(UInt(numerator)) at line 89
  was the same narrowing pattern fixed elsewhere.

- Config/ByronGenesis.swift: startTime Int → Int64. POSIX timestamp;
  Int32 overflows in 2038 ("Y2038 problem"). Affects 32-bit Linux as
  well as watchOS, not just watchOS.

- Crypto/Bip32.swift: derive(index:) and helpers Int → UInt32 per
  BIP32 spec. Hardened derivation keys at 0x8000_0000+ are > Int32.max,
  so on 32-bit watchOS hardened derivation traps on the `(1 << 31)`
  shift today. Replaced with `0x8000_0000` constant and explicit
  UInt32 arithmetic; the bit-or for hardened-bit setting is now
  obvious in intent.

+++ Sources/SwiftCardanoCore/Config/ByronGenesis.swift
@@ -10,8 +10,8 @@ public struct ByronGenesis: JSONLoadable {
     public let ftsSeed: String?
     /// Protocol constants
     public let protocolConsts: ProtocolConsts
-    /// Genesis start time
-    public let startTime: Int
+    /// Genesis start time (POSIX timestamp; Int64 to avoid Y2038 overflow on 32-bit platforms)
+    public let startTime: Int64
     /// Boot stakeholders
     public let bootStakeholders: [String: Int]
     /// Heavy delegation certificates
@@ -26,7 +26,7 @@ public struct ByronGenesis: JSONLoadable {
         blockVersionData: BlockVersionData,
         ftsSeed: String?,
         protocolConsts: ProtocolConsts,
-        startTime: Int,
+        startTime: Int64,
         bootStakeholders: [String: Int],
         heavyDelegation: [String: HeavyDelegation],
         nonAvvmBalances: [String: String],
